backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
bleach==6.0.0
blinker==1.6.2
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
ConfigArgParse==1.5.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
Django==4.2.4
djangorestframework==3.14.0
executing==1.2.0
filelock==3.12.2
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
fonttools==4.42.0
frozenlist==1.4.0
gevent==22.10.2
geventhttpclient==2.0.9
greenlet==2.0.2
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kaggle==1.5.15
kiwisolver==1.4.4
locust==2.15.1
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
msgpack==1.0.5
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.25.1
openai==0.27.8
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.5.3
plotly==5.16.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
pywin32==306
pyzmq==25.1.0
requests==2.31.0
roundrobin==0.0.4
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sklearn==0.0.post5
sqlparse==0.4.4
stack-data==0.6.2
tenacity==8.2.3
text-unidecode==1.3
threadpoolctl==3.1.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.23.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.6
yarl==1.9.2
zope.event==4.6
zope.interface==6.0
